backend/chat/requirements.txt
openai>=1.0.0
psycopg2-binary